+++ CHANGES.md
@@ -4,7 +4,8 @@
 
 - New xcube multi-level dataset rules:
   - `ml-dataset-meta`: verifies that a meta info file exists and is consistent
-  - `ml-dataset-resolution`: verifies that the levels have expected resolutions
+  - `ml-dataset-xy`: verifies that the levels have expected spatial resolutions
+  - `ml-dataset-time`: verifies that the levels have expected time dimension, if any
 - Now supporting xcube multi-level datasets `*.levels`:
   - Added xcube plugin processor `"xcube/multi-level-dataset"` that is used
     inside the predefined xcube configurations "all" and "recommended".